[Cleanup] Prevent float to double promotion
May reduce code slightly, but at least it makes execution faster.

+++ src/_P066_VEML6040.ino
@@ -145,9 +145,9 @@ boolean Plugin_066(uint8_t function, struct EventStruct *event, String& string)
         }
         case 2:
         {
-          UserVar[event->BaseVarIndex + 0] = pow(Plugin_066_CalcRelW(R, W), 0.4545) * 100.0f;
-          UserVar[event->BaseVarIndex + 1] = pow(Plugin_066_CalcRelW(G, W), 0.4545) * 100.0f;
-          UserVar[event->BaseVarIndex + 2] = pow(Plugin_066_CalcRelW(B, W), 0.4545) * 100.0f;
+          UserVar[event->BaseVarIndex + 0] = powf(Plugin_066_CalcRelW(R, W), 0.4545) * 100.0f;
+          UserVar[event->BaseVarIndex + 1] = powf(Plugin_066_CalcRelW(G, W), 0.4545) * 100.0f;
+          UserVar[event->BaseVarIndex + 2] = powf(Plugin_066_CalcRelW(B, W), 0.4545) * 100.0f;
           UserVar[event->BaseVarIndex + 3] = W;
           break;
         }
@@ -222,7 +222,7 @@ float Plugin_066_CalcCCT(float R, float G, float B)
   }
 
   float CCTi = (R - B) / G + 0.5f;
-  float CCT  = 4278.6f * pow(CCTi, -1.2455f);
+  float CCT  = 4278.6f * powf(CCTi, -1.2455f);
 
   return CCT;
 }
